feat(ontology): A1 — promote ConstraintEvaluation to Actions namespace
Move ConstraintEvaluation from Strategos.Ontology.Query to Strategos.Ontology.Actions;
update ActionConstraintReport and OntologyQueryService usings for backward compat.
TypeForwardedTo does not apply within a single assembly; Query consumers use explicit usings.
